requirements.txt
Django==1.6.5
Pillow==2.5.1
argparse==1.2.1
distribute==0.6.24
django-allauth==0.17.0
django-autocomplete-light==2.0.0a18
django-form-utils==1.0.1
oauthlib==0.6.3
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
